+++ client/src/css/ResourcesFilter.css
@@ -107,3 +107,7 @@ button {
   font-weight: 450;
   padding-left: 1.2em;
 }
+
+.resources-filter-autofill {
+  margin-left: 5%;
+}
